Putting a build copy of CPO in, and referencing it
Still working on the appropriate docs here, and this CPO stuff should get built
in. Unfortunately it's difficult right now to use CPO as an npm package
(because of CPO's somewhat unfriendly-to-npm build).

So we put a built copy source control for now. The dist/build/web directory has
everything that's needed to serve CPO, so as long as you make iframes that
point (however they are hosted) at dist/build/web/editor.embed.html, you should
get a working CPO page without having to reference out to code.pyret.org or
pyret-horizon.